requirements.txt
streamlit
pandas
numpy
altair
plotly
Pillow
